Converted hunger to Int
- Changed hunger to be an int value, now out of a max of 12
- Updated all places where hunger was used from an int to a float
- Updated cards and SOs to match
- Updated campfire

+++ SocialDeductionGame/Assets/Scripts/Events/Event Cards/DamageEvent.cs
@@ -8,7 +8,7 @@ public class DamageEvent : NightEvent
     [Header("Damage Ammount")]
     [SerializeField] private int _dmg;
     [Header("Hunger Loss Ammount")]
-    [SerializeField] private float _hunger;
+    [SerializeField] private int _hunger;
 
     // ========== METHOD OVERRIDES ==========
     public override void InvokeEvent()
